Move fixtures to conftest
